feat(vqb): snowflake-format-date tests moved to query transformation tests
